Move validator InterfaceHolder to dataactcore.
This is the 2nd step to combine the broker and validation InterfaceHolder
classes into a single class. The changeset moves all validation-specific
database logic from the specific validation interfaces (error, job,
staging, and validation) to the interfaces already in core. Thus,
those methods will be included when the consolidated InterfaceHolder
instantiates the broker-specific interfaces, because those are
sub-classes of the core interfaces. Again, this is all temporary
because the interfaces and their holders are deprecated.
